requirements.txt
fastapi==0.100.1
pydantic==2.7.1
python-multipart
Pillow
torch==2.5.1
torchvision==0.20.1
boto3
pinecone-client==4.1.0
openai-clip
ffmpeg
uvicorn[standard]
piexif


